+++ sync-for-expenses/CodatSyncExpenses/Models/Operations/DeleteConnectionResponse.cs
@@ -35,6 +35,6 @@ namespace Codat.Sync.Expenses.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }