+++ sync-for-expenses/CodatSyncExpenses/Models/Operations/UpdateExpenseTransactionResponse.cs
@@ -35,7 +35,7 @@ namespace Codat.Sync.Expenses.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
 
         /// <summary>
         /// Accepted