+++ sync-for-expenses/CodatSyncExpenses/Models/Webhooks/SyncStartedResponse.cs
@@ -29,6 +29,6 @@ namespace Codat.Sync.Expenses.Models.Webhooks
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }