+++ sync-for-expenses/RELEASES.md
@@ -131,3 +131,13 @@ Based on:
 - [csharp v4.0.0] sync-for-expenses
 ### Releases
 - [NuGet v4.0.0] https://www.nuget.org/packages/Codat.Sync.Expenses/4.0.0 - sync-for-expenses
+
+## 2023-11-20 01:29:57
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.0.1] sync-for-expenses
+### Releases
+- [NuGet v4.0.1] https://www.nuget.org/packages/Codat.Sync.Expenses/4.0.1 - sync-for-expenses